+++ app/layout.tsx
@@ -5,7 +5,7 @@ import './globals.css'
 import Header from '@/components/shared/header'
 import ActiveSectionContextProvider from '@/components/shared/activeSectionContextProvider'
 
-const inter = Inter({ subsets: ['latin'] })
+const inter = Inter({ subsets: ['latin'] });
 
 export const metadata: Metadata = {
   title: 'Shaik Shahbaaz Alam | Portfolio v4 | 2024',